Refactoring: temporarily splitting methods out to keep track
